refactor(widget): unificate widget config
